app/requirements.txt
Flask==2.3.3
redis==5.0.1
psycopg2-binary==2.9.7
prometheus-flask-exporter==0.23.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0